requirements.txt
discord~=1.7.3
python-dotenv~=0.20.0
spotipy~=2.19.0